chore: moved colors to /theme and export gitcoin-ui/theme
