+++ src/docs/colors.mdx
@@ -1,6 +1,6 @@
 import { Meta, ColorPalette, ColorItem, Title } from "@storybook/blocks";
 
-import { grantsStackUIKit, grantsStackDesignSystem } from "@/tokens/colors";
+import { grantsStackUIKit, grantsStackDesignSystem } from "@/theme/colors";
 
 <Meta title="Styles/Colors" />
